this commit deals with analytics views.py make a temp dir for each file when it is loaded to analytics:file_home save a limited number of figures for faster loading delete figure when number of files exceeds limit

+++ Element_Analytics/apps/analytics/views.py
@@ -20,11 +20,17 @@ current_file = None
 current_file_name = None
 current_frame = None
 headers = None
+temp_dir = None
+path = None
+MAX_TEMP_FIG = 5
 
 def file_home(request, file_name):
-    global current_file, current_frame, current_file_name, headers
+    global current_file, current_frame, current_file_name, headers, temp_dir, path
     current_file_name = file_name
     path = os.path.join(MEDIA_URL, 'documents/'+file_name)
+    temp_dir = os.path.join(MEDIA_URL, 'temp/'+file_name+'/')
+    if not os.path.exists(temp_dir):
+        os.makedirs(temp_dir)
     if not path.endswith(('.csv', '.log')):
         return HttpResponse("incorrect format")
     current_file = path
@@ -40,11 +46,19 @@ def variable_plot(request, file_name):
     if request.method == 'POST':
         col_name = request.POST['value']
 
-        # search if temp still has the col_name image
+        fig_path = os.path.join(temp_dir, col_name + '.png')
+        if not os.path.exists(fig_path):
+            gr = sb.distplot(current_frame[col_name]).get_figure()
+            gr.savefig(fig_path, format="png")
+            plt.close()
+
+        # if number of temp files > MAX_TEMP_FIG, remove earliest
+        temp_files = [os.path.join(temp_dir, f) for f in os.listdir(temp_dir)]
+        fig_files = [f for f in temp_files if f.endswith('.png') == True]
+        print(fig_files)
+        if len(fig_files) > MAX_TEMP_FIG:
+            _, f = min([(os.path.getctime(f), f) for f in fig_files])
+            os.remove(f)
         print(header)
-        gr = sb.distplot(current_frame[col_name]).get_figure()
-        fig_path = os.path.join(MEDIA_URL, 'temp/' + col_name+'.png')
-        gr.savefig(fig_path, format="png")
-        plt.close()
 
     return HttpResponse(open(fig_path, 'rb').read(), content_type="image/png")
